Write clients and docs to this repo
Outputs client code and docs to language dirs in dist/. The plan is for the
language dirs in dist to replace the computeclient_py and computeclient_js
repos. In addition to the generated code and docs, they'll also contain
non-generated content and configuration files for packaging and, hopefully,
tests too!